requirements.txt
python-telegram-bot[job-queue]==20.7
python-dotenv==1.0.0
Pillow==10.1.0
redis==5.0.1
aiofiles==23.2.1
aiosqlite==0.19.0
pandas==2.1.4
openpyxl==3.1.2
APScheduler==3.10.4
python-dateutil==2.8.2
pytz==2023.3